Merge branch 'phy_power'
Sebastian Hesselbarth says:

====================
net: phy: Ethernet PHY powerdown optimization

This is v2 of the ethernet PHY power optimization patches to reduce
power consumption of network PHYs with link that are either unused or
the corresponding netdev is down.

Compared to the last version, this patch set drops a patch to disable
unused PHYs after late initcall, as it is not compatible with a modular
mdio bus [1]. I'll investigate different ways to have a modular mdio bus
driver get notified when driver loading is done.

Again, a branch with v2 applied to v3.13-rc2 can also be found at
https://github.com/shesselba/linux-dove.git topic/ethphy-power-v2

[1] http://www.spinics.net/lists/arm-kernel/msg293028.html
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: b80b376c4451cc0d671a2b2f2191ea3f9857bd9f ]