Merge branch 'bnx2x'
Yuval Mintz says:

====================
bnx2x: Enhancements & semantic changes series

This patch series contains several semantic (or mostly semantic) patches,
as well as adding support for packet aggregations on the receive path
of windows VMs and updating bnx2x to the new FW recently accepted upstream.

Please consider applying these patches to `net-next'.

(This is a repost as net-next was still closed when this was previously
sent)
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 7282ec8cb4f82287c2aef20a24328b42c630f7cb ]