Merge branch 'mpls_tc'
Simon Wunderlich says:

====================
this series contains a header file proposal for MPLS labels. These
labels do not seem to be properly defined in the kernel so far. We are
developing a wired/wireless 802.21/MPLS switch and need to check the
MPLS labels to use the traffic control info for transmissions over
802.11 networks.

Changes to third version:

 * rename mpls_label_stack to mpls_label (thanks Neil)
 * fix over-indendented closing brac (thanks Sergei)
 * add Johannes' Ack
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 9e82e7f4af0ad90ab41b4ae80ce42b20df20c725 ]