Merge branch 'tipc'
Jon Maloy says:

====================
tipc: simplifications in socket and port layer

After the removal of the tipc native API the relation between
tipc_port and its API types is strictly one-to-one, i.e, the latter
can now only be a socket API. This change opens up for
simplifications both in the code, data and locking structure.

We start with this series, where we ensure that port and socket
structures are co-allocated. Note that the first commit in the
series is unrelated to the above.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 36f6fdb74997c3f456f1c6a9803c009c49a30053 ]