Merge branch 'dev_kfree_skb_any'
Eric W. Biederman says:

====================
Using dev_kfree_skb_any for functions called in multiple contexts

This patchset should be an uncontroversial set of changes to change
dev_kfree_skb to dev_kfree_skb_any for code paths that are called in
hard irq contexts in addition to other contexts.  netpoll is the reason
this code gets called in multiple contexts.

There is more coming but these changes are a good starting place, and
stand on their own.

Since the last round changes to the rx path have been removed netpoll
will changed to avoid that.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>
Acked-by: Eric Dumazet <edumazet@google.com>

[ upstream commit: 8a83d9396a4945dcaa129b14a725311e5846b2d5 ]