ARM: ux500: Add a vmmc regulator through DT for the poped eMMC for href
Add a node in DT for the proper regulator which means we can move away
from the mmci platform data which currently holds the corresponding OCR
mask.

The mmci driver can then calculate the OCR mask based on the voltages
supported by the regulator, instead of relying on the platform data.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Acked-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: e1a309a4fa77d265afbc1158f0f4ec48752da6a6 ]