Merge branch 'tipc-next'
Erik Hugne says:

====================
tipc: fix handling of NETDEV_CHANGEADDR event

Aside from manual reconfiguration of the netdevice hwaddr, this can also
be changed automatically for an interface bond in active-backup mode
if fail_over_mac is enabled. This patchset fixes the handling of this
event in TIPC by properly updating the l2 media address for the bearer,
followed by a reinitialization of the node discovery mechanism.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: f3c9ec05ab020b57108647cd30ce67bec156486f ]