Merge branch 'net_sysfs_doc'
Florian Fainelli says:

====================
net: document sysfs entries

This patchset attempts to document the basic set of sysfs entries that are
exposed by netdevices in /sys/class/net/<iface>/

I did not go before the pre-git era, so the oldest entries are marked with
the 2.6.12 kernel version and dated of April 2005.

Future patches will document the queues/ and statistics/ directories as well.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 788367fdb30329d9db92c89a7b9ca1a0cbea5c36 ]